estudo_de_caso/requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.22
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0
scikit-learn==1.3.2
streamlit==1.29.0
dash==2.14.2
bokeh==3.3.2
altair==5.2.0
python-dotenv==1.0.0